pyproject.toml
[project]
name = "rose-py"
version = "0.1.5"
description = "A lightweight ROS2-like framework based on Zenoh — no Ubuntu lock-in, no compilation needed, no .msg files.  一个类似 ROS 的基于 Zenoh 的轻量级 发布/订阅 与 RPC 通信框架, 不挑平台、无需编译、没有 .msg"
readme = "README.md"
authors = [
    { name = "DBinK", email = "DBinKv1@Gmail.com" }
]
requires-python = ">=3.10"
dependencies = [
    "eclipse-zenoh>=1.9.0",
    "msgspec>=0.21.1",
    "typer>=0.25.1",
    "loguru>=0.7.3",
    "rich>=15.0.0",
]

[project.scripts]
rrr = "rose.probe:app"

[project.urls]
"repository" = "https://github.com/DBinK/rose"

[build-system]
requires = ["uv_build>=0.9.21,<0.10.0"]
build-backend = "uv_build"

[tool.uv.build-backend]
module-name = "rose"
module-root = "src"

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-benchmark>=5.1.0",
    "opencv-python-headless>=4.10.0",
]

[[tool.uv.index]]
url = "https://pypi.org/simple"
# url = "https://pypi.tuna.tsinghua.edu.cn/simple"  # 清华源
default = true

[tool.pyrefly]
python-path = ".venv/Scripts/python.exe"
